Use najax when DOM is not present
When the DOM isn't specified we can assume that this service is being
used within the context of Fastboot.  Unfortunately, Ember.$.ajax won't
do the correct thing.  Fastboot has created a service:

https://github.com/tildeio/ember-cli-fastboot/blob/master/app/initializers/server/ajax.js

That implements adapter#ajax in the context of node using `najax`

This checks to see if the DOM is present and if so behaves as it did
before, however when DOM isn't present it will assume Fastboot has
loaded najax and will call it passsing in the same `hash` parameters