Create AJAX service from ES6 class
Uses a function that takes an ES6 class and returns a correctly-working Ember mixin.  By mixin the ES6 class into the service, we can avoid needing to duplicate all of that code.  All tests still pass, completely un-modified.  Should work with IE9+ but not lower, since the mixin-creating function relies on an HTML5 API that isn't available in IE versions lower than that.