Remove duplicate service tests
Removes the original set of tests that were ported over to the AjaxRequest class, so that we don't have duplciates of all of them that need to be maintained.  And verifification that the Ember version of the class behaves the same as the ES6 version should really happen in the `ember-mixinify-class` tests instead.