Make the header tests more readable
Previous, it was a little hard to tell what exactly the header tests were doing.  They also depended on the presence of a `.beforeSend` method on the configuration object, which wasn't guaranteed to exist if the headers are not set on the request.

Additionally, this adds a second helper method for returning JSON data from Pretender, and refactors the json test helper to export both functions instead of a default, since neither made more sense to be the default one.  The rest of the changes just import the json factory helper correctly.