Allow an absolute URL to override host settings
Previously, if a request was constructed with a host in the URL itself, it would create a combination url that looked something like this:

```
request('http://foo.com', { host: 'http://bar.com' })
// hits the url "http://bar.com/http://foo.com"
```

The desired outcome in these cases is probably to instead override the host in the configuration, which is now what happens instead.

This also adds a handle URL parsing class that, given a string URL, will populate itself with all of the individual pieces of the URL broken down.  It has been constructed in a way that should allow it continue to work once this is running in a Node.js platform as well.