Updated URLs in options tests
The usage of the `request` method in the tests didn't match how it
should actually be used, and that made itself clear in the fact that the
namespace-handling functionality broke these tests.  Since the tests
showed `example.com` as being a valid argument for the request URL, one
might think that that's how the addon should be used, when in reality a
URL used there that does not include the full protocol will not be used
as they might expect, if they also have a host defined.  Therefore, to
fix the test cases and make them more clear, I've updated the tests to
reflect the expected usage.