Change `isAbsolute` property to `isComplete`
The check was not actually whether or not a URL was absolute, but
whether it was "complete".  For example, "http://foo.com" should have
been recognized, although that is not actually what an "absolute url"
usually refers to, which is something like "/foo".  The name change just
increases the clarify of what is being verifyied.