Prevent removing trailing slash in request URL (#96)
* Change `isAbsolute` property to `isComplete`

The check was not actually whether or not a URL was absolute, but
whether it was "complete".  For example, "http://foo.com" should have
been recognized, although that is not actually what an "absolute url"
usually refers to, which is something like "/foo".  The name change just
increases the clarify of what is being verifyied.

* Correctly build full URL from parts

The old way of building a URL was too brittle and didn't cover a lot of
the different cases that could take place, in different combinations of
the host, namespace, and segment having or not having slashes included.

The new approach implemented here handles the various cases correctly,
and ensures that we do not remove a trailing slash if one was provided.

Closes #70